Function pallet_pool_registry::pallet::dispatchables::register

pub fn register<T: Config>(
    admin: T::AccountId,
    pool_id: T::PoolId,
    tranche_inputs: Vec<<<T as Config>::ModifyPool as PoolMutate<<T as Config>::AccountId, <T as Config>::PoolId>>::TrancheInput>,
    currency: T::CurrencyId,
    max_reserve: T::Balance,
    metadata: Option<Vec<u8>>,
    write_off_policy: <<T as Config>::ModifyWriteOffPolicy as PoolWriteOffPolicyMutate<<T as Config>::PoolId>>::Policy
)

Create a new pool

Initialise a new pool with the given ID and tranche configuration. Tranche 0 is the equity tranche, and must have zero interest and a zero risk buffer.

The minimum epoch length, and maximum NAV age will be set to chain-wide defaults. They can be updated with a call to update.

The caller will be given the PoolAdmin role for the created pool. Additional administrators can be added with the Permissions pallet.

Returns an error if the requested pool ID is already in use, or if the tranche configuration cannot be used.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the register variant of the pallet Call enum. You should not attempt to call this function directly.