Function pallet_pool_registry::pallet::dispatchables::update

pub fn update<T: Config>(
    pool_id: T::PoolId,
    changes: <<T as Config>::ModifyPool as PoolMutate<<T as Config>::AccountId, <T as Config>::PoolId>>::PoolChanges
)

Update per-pool configuration settings.

This updates the tranches of the pool, sets the minimum epoch length, and maximum NAV age.

If no delay is required for updates and redemptions don’t have to be fulfilled, then this is executed immediately. Otherwise, the update is scheduled to be executed later.

The caller must have the PoolAdmin role in order to invoke this extrinsic.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the update variant of the pallet Call enum. You should not attempt to call this function directly.