Trait pallet_pool_registry::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>;
    type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen + Debug;
    type InterestRate: Member + Parameter + Default + Copy + TypeInfo + FixedPointNumber<Inner = Self::Balance>;
    type ModifyPool: PoolMutate<Self::AccountId, Self::PoolId, CurrencyId = Self::CurrencyId, Balance = Self::Balance>;
    type ModifyWriteOffPolicy: PoolWriteOffPolicyMutate<Self::PoolId>;
    type TrancheCurrency: TrancheCurrency<Self::PoolId, Self::TrancheId> + Into<Self::CurrencyId>;
    type CurrencyId: Parameter + Copy;
    type TrancheId: Member + Parameter + Default + Copy + MaxEncodedLen + TypeInfo + From<[u8; 16]>;
    type PoolCreateOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type MaxTokenNameLength: Get<u32> + Copy + Member + TypeInfo;
    type MaxTokenSymbolLength: Get<u32> + Copy + Member + TypeInfo;
    type MaxTranches: Get<u32> + Member + TypeInfo;
    type MaxSizeMetadata: Get<u32> + Copy + Member + TypeInfo;
    type Permission: Permissions<Self::AccountId, Scope = PermissionScope<Self::PoolId, Self::CurrencyId>, Role = Role<Self::TrancheId>, Error = DispatchError>;
    type AssetRegistry: Mutate<AssetId = Self::CurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata> + Inspect<AssetId = Self::CurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>

type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen + Debug

type InterestRate: Member + Parameter + Default + Copy + TypeInfo + FixedPointNumber<Inner = Self::Balance>

A fixed-point number which represents an interest rate.

type ModifyPool: PoolMutate<Self::AccountId, Self::PoolId, CurrencyId = Self::CurrencyId, Balance = Self::Balance>

type ModifyWriteOffPolicy: PoolWriteOffPolicyMutate<Self::PoolId>

type TrancheCurrency: TrancheCurrency<Self::PoolId, Self::TrancheId> + Into<Self::CurrencyId>

The currency type of investments.

type CurrencyId: Parameter + Copy

type TrancheId: Member + Parameter + Default + Copy + MaxEncodedLen + TypeInfo + From<[u8; 16]>

type PoolCreateOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin permitted to create pools

type MaxTokenNameLength: Get<u32> + Copy + Member + TypeInfo

Max length for a tranche token name

type MaxTokenSymbolLength: Get<u32> + Copy + Member + TypeInfo

Max length for a tranche token symbol

type MaxTranches: Get<u32> + Member + TypeInfo

Max number of Tranches

type MaxSizeMetadata: Get<u32> + Copy + Member + TypeInfo

Max size of Metadata

type Permission: Permissions<Self::AccountId, Scope = PermissionScope<Self::PoolId, Self::CurrencyId>, Role = Role<Self::TrancheId>, Error = DispatchError>

type AssetRegistry: Mutate<AssetId = Self::CurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata> + Inspect<AssetId = Self::CurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata>

The registry type used for retrieving and updating tranche metadata as part of the PoolMetadata trait implementation

type WeightInfo: WeightInfo

Weight Information